DBMS of Criminal Records for Seattle Police Department
Created a DBMS to address the business requirements for Seattle Crime
Records to keep track of criminal records of suspects who have been
acquitted and sentenced to imprisonment. The main purpose of this
database is to facilitate efficient and quick retrieval of prisoner
information for Seattle prison administration. It focuses mainly on the
prisoner's imprisonment in the facilities under the department’s
jurisdiction. The database is designed for Seattle police department to
pull data on the criminals currently serving prison time. Being able to
access the data quickly will help police department run administration
smoothly. Additionally, the police department will be able to analyze
data about criminal’s health, employment, family, parole, visitors etc
to collate and maintain in depth information for future criminal
investigation.